requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
streamlit>=1.24.0
openai>=1.3.0

# For Python 3.13, we need to install these separately with pre-built wheels
# Run these commands manually if the script fails:
# pip install --only-binary :all: numpy pandas ta
duckduckgo-search>=5.0.0
beautifulsoup4>=4.12.0
